main.c norm
